multithreading / now can handle simul downloads
